Improve layout responsiveness and enhance content visibility
Adjusts the layout of the ContactPage for better alignment in smaller screens by changing the flex alignment.

Additionally, adds a conditional rendering for title and description on the PortfolioPage for desktop views, enhancing user experience by providing more context.

These changes aim to improve the overall usability and aesthetics of the pages.